Developed modelPhaseSpace & defineVariableValues
We have a working model! But it has issues, the chirp is somehow
reversed, (and so had to be re-reversed)
and the scaling doesn't make sense (values from the function at a point
have to be multipled by 10^14 instead of 10^12 and they don't increase
as much as they should). A possibility is that issues arrise from how
far the decimal places go in the exp and pow functions, or that the
formula has a written error.