fixed value issue
problem arose from accuracy. the height range is 1005 long from ~-500 to 500. So once the phase space was divided more than 1005, the distance became smaller than 1, so the accuracy became inaccurate because multiplied with the intensityRatio it represents that square of intensity, accuracy by accuracy wide and tall. However when the distance becomes say 0.5 by 0.5, however the intensity gathered is multiplied still by 1 it becomes 4x what it should be as it represents a square 1x1.